CARF-545 Restore cleanup in RegistrationTypePage but clear match flag separately in controller if needed (scenario 2b)
