+++ app/models/OrganisationRegistrationType.scala
@@ -17,7 +17,6 @@
 package models
 
 import play.api.i18n.Messages
-import play.api.libs.json.{JsString, Writes}
 import uk.gov.hmrc.govukfrontend.views.Aliases.Text
 import uk.gov.hmrc.govukfrontend.views.viewmodels.radios.RadioItem
